MSEC-36: [SharedKernel] Create a Core poject
