+++ src/MeraStore.Shared.Common.Core/Domain/Entities/Entity.cs
@@ -1,0 +1,6 @@
+﻿namespace MeraStore.Shared.Common.Core.Domain.Entities;
+
+public abstract class Entity
+{
+    public Guid Id { get; protected set; } = Guid.NewGuid();
+}
