+++ src/MeraStore.Shared.Common.Core/Enums/Currency.cs
@@ -1,0 +1,16 @@
+﻿namespace MeraStore.Shared.Common.Core.Enums;
+
+public enum Currency
+{
+  INR,  // Indian Rupee
+  USD,  // United States Dollar
+  EUR,  // Euro
+  GBP,  // British Pound Sterling
+  JPY,  // Japanese Yen
+  AUD,  // Australian Dollar
+  CAD,  // Canadian Dollar
+  CHF,  // Swiss Franc
+  CNY,  // Chinese Yuan
+  SEK,  // Swedish Krona
+  NZD,  // New Zealand Dollar
+}
