+++ src/MeraStore.Shared.Common.Core/ErrorsCodes/ErrorCodeProvider.cs
@@ -1,4 +1,4 @@
-﻿namespace MeraStore.Shared.Common.ErrorsCodes;
+﻿namespace MeraStore.Shared.Common.Core.ErrorsCodes;
 
 public class ErrorCodeProvider
 {
