+++ src/MeraStore.Shared.Common.Core/ErrorsCodes/ServiceProvider.cs
@@ -1,4 +1,4 @@
-﻿namespace MeraStore.Shared.Common.ErrorsCodes;
+﻿namespace MeraStore.Shared.Common.Core.ErrorsCodes;
 
 public class ServiceProvider
 {
