+++ src/MeraStore.Shared.Common.Core/Events/IEventHandler.cs
@@ -1,0 +1,6 @@
+﻿namespace MeraStore.Shared.Common.Core.Events;
+
+public interface IEventHandler<TEvent> where TEvent : DomainEventBase
+{
+  Task Handle(TEvent domainEvent);
+}
