+++ src/MeraStore.Shared.Common.Core/Events/IIntegrationEventHandler.cs
@@ -1,0 +1,6 @@
+﻿namespace MeraStore.Shared.Common.Core.Events;
+
+public interface IIntegrationEventHandler<TEvent> where TEvent : IntegrationEvent
+{
+  Task Handle(TEvent integrationEvent);
+}
