+++ src/MeraStore.Shared.Common.Core/Events/IntegrationEvent.cs
@@ -1,0 +1,6 @@
+﻿namespace MeraStore.Shared.Common.Core.Events;
+
+public abstract class IntegrationEvent : EventBase
+{
+  
+}
