+++ src/MeraStore.Shared.Common.Core/Exceptions/InventoryServiceExceptions.cs
@@ -1,0 +1,19 @@
+﻿using System.Net;
+using MeraStore.Shared.Common.Core.ErrorsCodes;
+
+namespace MeraStore.Shared.Common.Core.Exceptions;
+
+public class InventoryServiceExceptions
+{
+    public class OutOfStockException(string message) : BaseAppException(
+      ServiceProvider.GetServiceCode(Constants.ServiceIdentifiers.InventoryService),
+      EventCodeProvider.GetEventCode(Constants.EventCodes.ResourceNotFound),
+      ErrorCodeProvider.GetErrorCode(Constants.ErrorCodes.NotFoundError),
+      HttpStatusCode.NotFound, message);
+
+    public class StockAdjustmentException(string message) : BaseAppException(
+      ServiceProvider.GetServiceCode(Constants.ServiceIdentifiers.InventoryService),
+      EventCodeProvider.GetEventCode(Constants.EventCodes.ServiceError),
+      ErrorCodeProvider.GetErrorCode(Constants.ErrorCodes.UnprocessableEntityError),
+      HttpStatusCode.UnprocessableEntity, message);
+}
