+++ src/MeraStore.Shared.Common.WebApi/Middlewares/CorrelationIdMiddleware.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using MeraStore.Shared.Common.Core;
+using Microsoft.AspNetCore.Http;
 using Serilog.Context;
 
 namespace MeraStore.Shared.Common.WebApi.Middlewares;
