--- src/MeraStore.Shared.Common/Exceptions/CartServiceExceptions.cs
@@ -1,19 +1,0 @@
-﻿using System.Net;
-using MeraStore.Shared.Common.ErrorsCodes;
-
-namespace MeraStore.Shared.Common.Exceptions;
-
-public class CartServiceExceptions
-{
-  public class CartNotFoundException(string message) : BaseAppException(
-    ServiceProvider.GetServiceCode(Constants.ServiceIdentifiers.CartService),
-    EventCodeProvider.GetEventCode(Constants.EventCodes.ResourceNotFound),
-    ErrorCodeProvider.GetErrorCode(Constants.ErrorCodes.NotFoundError),
-    HttpStatusCode.NotFound, message);
-
-  public class CartItemNotFoundException(string message) : BaseAppException(
-    ServiceProvider.GetServiceCode(Constants.ServiceIdentifiers.CartService),
-    EventCodeProvider.GetEventCode(Constants.EventCodes.ResourceNotFound),
-    ErrorCodeProvider.GetErrorCode(Constants.ErrorCodes.NotFoundError),
-    HttpStatusCode.NotFound, message);
-}
